// src/app/features/cart/models/cart.interface.ts
export interface Cart {
  _id: string;
  cartOwner: string;
  products: Product2[];
  createdAt: string;
  updatedAt: string;
  __v: number;
  totalCartPrice: number;
}

export interface CartResponse {
  status: string;
  numOfCartItems?: number;
  cartId?: string;
  message?: string;
  data: Cart;
}
export interface Product2 {
  count: number;
  _id: string;
  product: Product;
  price: number;
}
export interface Product {
  subcategory: Subcategory[];
  _id: string;
  title: string;
  slug?: string;
  quantity: number;
  imageCover: string;
  category: Category;
  brand: Category;
  ratingsAverage: number;
  id: string;
}
export interface Category {
  _id: string;
  name: string;
  slug: string;
  image: string;
}

export interface Subcategory {
  _id: string;
  name: string;
  slug: string;
  category: string;
}